requirements.txt
Django==1.9.5
django-pipeline==1.6.8
djangorestframework==3.3.3
Markdown==2.6.6
Pygments==2.1.3
